+++ lib/src/analyzer/rules/registry.dart
@@ -1,5 +1,6 @@
 import 'package:analysis_server_plugin/registry.dart';
 
+import 'custom_hooks_require_build_rule.dart';
 import 'disallow_context_outside_build_rule.dart';
 import 'discarded_global_effect_rule.dart';
 import 'effect_cleanup_requires_effect_rule.dart';
@@ -12,6 +13,7 @@ import 'scope_dispose_requires_scope_rule.dart';
 void registerRules(PluginRegistry registry) {
   registry.registerWarningRule(RequireContextInBuildRule());
   registry.registerWarningRule(DisallowContextOutsideBuildRule());
+  registry.registerWarningRule(CustomHooksRequireBuildRule());
   registry.registerWarningRule(NoHooksInControlFlowRule());
   registry.registerWarningRule(NoHooksInNestedFunctionsRule());
   registry.registerWarningRule(NoWritesInComputedRule());